+++ SmartButtons/SmartButtons/View Model/LearnQuizViewModel.swift
@@ -8,15 +8,15 @@
 import Foundation
 // class extends protocol
 final class LearnQuizViewModel: LearnQuiz {
-    
+
     func multiply(_ number1: Int, _ number2: Int) -> Int {
         let ans = number1 * number2
         return ans
     }
     
-    func division(_ number1: Int, _ number2: Int) -> Int {
+    func division(_ number1: Int, _ number2: Int) -> Double {
        let ans = number2/number1
-        return ans
+        return Double(ans)
     }
     
 }
